debt: Moved replication redux bits into their own duck
