+++ src/drive/mobile/ducks/mediaBackup/styles.styl
@@ -1,5 +1,4 @@
 @require 'settings/palette.styl'
-// TODO: WHY ???? @require '../../styles/feedback'
 
 .coz-upload-status-wrapper
     position relative
