+++ src/drive/mobile/ducks/settings/actions.js
@@ -1,15 +1,6 @@
-import { openFolder, getOpenedFolderId } from '../../../actions'
-
-import { startReplication as startPouchReplication } from '../../lib/replication'
 import { startTracker, stopTracker } from '../../lib/tracker'
-import { resetClient } from '../../lib/cozy-helper'
-import { revokeClient as reduxRevokeClient } from '../authorization'
 
-import {
-  getServerUrl,
-  isFirstReplicationDone,
-  getPouchIndexes
-} from './reducers'
+import { getServerUrl } from './reducers'
 
 export const SET_URL = 'SET_URL'
 export const BACKUP_IMAGES = 'BACKUP_IMAGES'
@@ -17,8 +8,6 @@ export const BACKUP_CONTACTS = 'BACKUP_CONTACTS'
 export const WIFI_ONLY = 'WIFI_ONLY'
 export const SET_ANALYTICS = 'SET_ANALYTICS'
 export const SET_OFFLINE = 'SET_OFFLINE'
-export const SET_FIRST_REPLICATION = 'SET_FIRST_REPLICATION'
-export const SET_POUCH_INDEXES = 'SET_POUCH_INDEXES'
 
 export const setUrl = url => ({ type: SET_URL, url })
 
@@ -39,46 +28,7 @@ export const setAnalytics = (analytics, source = 'settings') => (
   }
 }
 
-export const setFirstReplication = firstReplication => ({
-  type: SET_FIRST_REPLICATION,
-  firstReplication
-})
-
-export const setPouchIndexes = indexes => ({
-  type: SET_POUCH_INDEXES,
-  indexes
-})
-
 export const setBackupImages = backupImages => ({
   type: BACKUP_IMAGES,
   backupImages
 })
-
-export const startReplication = () => (dispatch, getState) => {
-  console.info('Starting replication...')
-
-  const firstReplication = isFirstReplicationDone(getState())
-  const existingIndexes = getPouchIndexes(getState())
-  const refreshFolder = () => {
-    dispatch(openFolder(getOpenedFolderId(getState())))
-  }
-  const revokeClient = () => {
-    resetClient()
-    dispatch(reduxRevokeClient())
-  }
-  const firstReplicationFinished = () => {
-    dispatch(setFirstReplication(true))
-  }
-  const indexesCreated = indexes => {
-    dispatch(setPouchIndexes(indexes))
-  }
-
-  startPouchReplication(
-    existingIndexes,
-    firstReplication,
-    firstReplicationFinished,
-    refreshFolder,
-    revokeClient,
-    indexesCreated
-  )
-}
