+++ src/drive/mobile/ducks/settings/reducers.js
@@ -1,8 +1,5 @@
-/* global __TARGET__ */
 import {
   SET_OFFLINE,
-  SET_FIRST_REPLICATION,
-  SET_POUCH_INDEXES,
   SET_URL,
   BACKUP_IMAGES,
   BACKUP_CONTACTS,
@@ -11,12 +8,10 @@ import {
 } from './actions'
 
 const initialState = {
-  offline: false,
-  firstReplication: false,
-  indexes: null,
   serverUrl: '',
-  backupImages: false,
+  offline: false,
   analytics: false,
+  backupImages: false,
   wifiOnly: true
 }
 
@@ -26,16 +21,12 @@ const settings = (state = initialState, action) => {
       return { ...state, serverUrl: action.url }
     case SET_OFFLINE:
       return { ...state, offline: action.offline }
-    case SET_FIRST_REPLICATION:
-      return { ...state, firstReplication: action.firstReplication }
-    case SET_POUCH_INDEXES:
-      return { ...state, indexes: action.indexes }
+    case SET_ANALYTICS:
+      return { ...state, analytics: action.analytics }
     case BACKUP_IMAGES:
       return { ...state, backupImages: action.backupImages }
     case BACKUP_CONTACTS:
       return { ...state, backupContacts: action.backupContacts }
-    case SET_ANALYTICS:
-      return { ...state, analytics: action.analytics }
     case WIFI_ONLY:
       return { ...state, wifiOnly: action.wifiOnly }
     default:
@@ -62,13 +53,3 @@ export const isOfflineCapable = state => getSetting(state, 'offline')
 export const isImagesBackupOn = state => getSetting(state, 'backupImages')
 export const isWifiOnlyOn = state => getSetting(state, 'wifiOnly')
 export const isAnalyticsOn = state => getSetting(state, 'analytics')
-export const getPouchIndexes = state => getSetting(state, 'indexes')
-
-export const isFirstReplicationDone = state =>
-  getSetting(state, 'firstReplication')
-
-export const shouldWorkFromPouchDB = state =>
-  __TARGET__ === 'mobile' &&
-  isOfflineCapable(state) &&
-  isFirstReplicationDone(state) &&
-  getPouchIndexes(state) !== null
