+++ src/drive/mobile/ducks/settings/styles.styl
@@ -1,4 +1,3 @@
-// @require 'cozy-ui'
 @require 'components/forms.styl'
 
 .settings
